requirements.txt
fastapi
uvicorn[standard]
pdfminer.six
pandas
python-multipart
requests